feat: extract two metods from media to twilio service
