+++ VoiceCallAssistant/Messages/BaseEvent.cs
@@ -1,0 +1,6 @@
+﻿namespace VoiceCallAssistant.Messages;
+
+public abstract class BaseEvent
+{
+    public string @Event { get; set; } = null!;
+}
