+++ VoiceCallAssistant/Services/TwilioService.cs
@@ -1,8 +1,14 @@
-﻿using Twilio;
+﻿using OpenAI.RealtimeConversation;
+using System.Collections.Concurrent;
+using System.Net.WebSockets;
+using System.Text;
+using System.Text.Json;
+using Twilio;
 using Twilio.Rest.Api.V2010.Account;
 using Twilio.TwiML;
 using Twilio.TwiML.Voice;
 using VoiceCallAssistant.Interfaces;
+using Task = System.Threading.Tasks.Task;
 
 namespace VoiceCallAssistant.Services;
 
@@ -74,5 +80,150 @@ public class TwilioService : ITwilioService
 
         Console.WriteLine($"Returning TwiML for the outbound call");
         return response.ToString();
+    }
+
+    public async Task ReceiveFrom(
+            WebSocket webSocket,
+            CancellationToken ct,
+            Action<string> setStreamSid,
+            Action<BinaryData, long> handleAudio,
+            ConcurrentQueue<string> markQueue)
+    {
+        var buffer = new byte[4 * 1024];
+
+        while (!ct.IsCancellationRequested && webSocket.State == WebSocketState.Open)
+        {
+            var result = await webSocket.ReceiveAsync(
+                new ArraySegment<byte>(buffer), ct);
+
+            if (result.MessageType == WebSocketMessageType.Close)
+            {
+                await webSocket.CloseAsync(
+                    WebSocketCloseStatus.NormalClosure,
+                    "closing", CancellationToken.None);
+                break;
+            }
+
+            var jsonString = Encoding.UTF8.GetString(buffer, 0, result.Count);
+            using var doc = JsonDocument.Parse(jsonString);
+            var root = doc.RootElement;
+            var evt = root.GetProperty("event").GetString();
+
+            switch (evt)
+            {
+                case "start":
+                    {
+                        var sid = ExtractStreamSid(root);
+
+                        setStreamSid(sid);
+                        break;
+                    }
+                case "media":
+                    {
+                        var (audioBinary, tsLong) = ExtractPayload(root);
+
+                        handleAudio(audioBinary, tsLong);
+                        break;
+                    }
+                case "mark":
+                    {
+                        markQueue.TryDequeue(out _);
+                        break;
+                    }
+                case "stop":
+                    {
+                        await webSocket.CloseAsync(
+                            WebSocketCloseStatus.NormalClosure,
+                            "closing", CancellationToken.None);
+                        break;
+                    }
+            }
+        }
+    }
+
+    public async Task SendTo(RealtimeConversationSession session,
+        CancellationToken ct,
+        Func<ConversationItemStreamingPartDeltaUpdate, Task> handleAudioDelta,
+        Func<ConversationInputSpeechStartedUpdate, Task> handleSpeechStarted)
+    {
+        await foreach (ConversationUpdate update in session.ReceiveUpdatesAsync(ct))
+        {
+            // Notification indicating the start of the conversation session.
+            if (update is ConversationSessionStartedUpdate sessionStartedUpdate)
+            {
+                // Start conversation first
+                await session.StartResponseAsync();
+
+                Console.WriteLine($"<<< Session started. ID: {sessionStartedUpdate.SessionId}");
+                Console.WriteLine();
+            }
+
+            // Notification indicating the start of detected voice activity.
+            if (update is ConversationInputSpeechStartedUpdate speechStartedUpdate)
+            {
+                Console.WriteLine(
+                    $"  -- Voice activity detection started at {speechStartedUpdate.AudioStartTime}");
+                await handleSpeechStarted(speechStartedUpdate);
+            }
+
+            // Notification indicating the end of detected voice activity.
+            if (update is ConversationInputSpeechFinishedUpdate speechFinishedUpdate)
+            {
+                Console.WriteLine(
+                    $"  -- Voice activity detection ended at {speechFinishedUpdate.AudioEndTime}");
+            }
+
+            // Notification about item streaming delta, which may include audio transcript, audio bytes, or function arguments.
+            if (update is ConversationItemStreamingPartDeltaUpdate deltaUpdate)
+            {
+
+                // Handle audio bytes.
+                if (deltaUpdate.AudioBytes is not null)
+                {
+                    await handleAudioDelta(deltaUpdate);
+                }
+            }
+
+            // Notification indicating the completion of transcription from input audio.
+            if (update is ConversationInputTranscriptionFinishedUpdate transcriptionCompletedUpdate)
+            {
+                Console.WriteLine();
+                Console.WriteLine($"  -- User audio transcript: {transcriptionCompletedUpdate.Transcript}");
+                Console.WriteLine();
+            }
+
+            // Notification about error in conversation session.
+            if (update is ConversationErrorUpdate errorUpdate)
+            {
+                Console.WriteLine();
+                Console.WriteLine($"ERROR: {errorUpdate.Message}");
+                break;
+            }
+        }
+    }
+
+    private (BinaryData audioBinary, long ts) ExtractPayload(JsonElement root)
+    {
+        var payloadB64 = root
+            .GetProperty("media")
+            .GetProperty("payload")
+            .GetString()!;
+        var ts = root
+            .GetProperty("media")
+            .GetProperty("timestamp")
+            .GetString();
+        var audioBytes = Convert.FromBase64String(payloadB64);
+        var audioBinary = new BinaryData(audioBytes);
+        var tsLong = Convert.ToInt64(ts);
+
+        return (audioBinary, tsLong);
+    }
+
+    private string ExtractStreamSid(JsonElement root)
+    {
+        return root
+            .GetProperty("start")
+            .GetProperty("streamSid")
+            .GetString()!;
     }
 }
